requirements.txt
Django==2.0.1
django-bootstrap4==0.0.6
Pillow==5.1.0
pytz==2018.3
